Change karma config files to `ts`
