--- config/karma/karma.conf.dev.cjs
@@ -1,9 +1,0 @@
-const setKarmaBaseConfig = require('../../karma.conf.cjs');
-
-module.exports = function (config) {
-    setKarmaBaseConfig(config);
-    const newConfig = {
-        singleRun: false
-    }
-    config.set(newConfig);
-}
