--- config/karma/karma.conf.prod.cjs
@@ -1,9 +1,0 @@
-const setKarmaBaseConfig = require('../../karma.conf.cjs');
-
-module.exports = function (config) {
-    setKarmaBaseConfig(config);
-    const newConfig = {
-        singleRun: true
-    }
-    config.set(newConfig);
-}
